requirements.txt
aiomysql==0.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
async-sqlalchemy==1.0.0
asyncpg==0.30.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dill==0.3.9
fastapi==0.115.5
filelock==3.16.1
fonttools==4.55.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pydantic==2.10.1
pydantic_core==2.27.1
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.37
ultralytics-thop==2.0.12
urllib3==2.2.3
uvicorn==0.32.1
pillow==10.0.1   